@@ -6,12 +6,12 @@
     "homepage": "https://github.com/aherne/php-servlets-api",
     "license": "MIT",
     "require": {
-        "php": "^7.1",
+        "php": "^8.1",
         "ext-simplexml": "*",
-        "lucinda/abstract_mvc": "^1.0"
+        "lucinda/abstract_mvc": "^2.0"
     },
     "require-dev": {
-    	"lucinda/unit-testing": "^1.0"
+    	"lucinda/unit-testing": "^2.0"
     },
     "autoload": {
     	"psr-4": {
