@@ -1,28 +1,68 @@
 <?php
 namespace Lucinda\STDOUT\Application;
 
+use Lucinda\MVC\ConfigurationException;
 use Lucinda\STDOUT\Application\Route\Parameter;
+use Lucinda\STDOUT\Request\Method;
 
 /**
- * Encapsulates route information:
- * - url: relative path requested
- * - cpntroller: path to controller (relative to application/controllers folder)
- * - view: path to view (relative to application/views folder)
+ * Encapsulates extra route information for request validation
  */
 class Route extends \Lucinda\MVC\Application\Route
 {
-    private $requestMethod;
-    private $parameters = [];
-    
+    private ?Method $requestMethod = null;
+    private array $parameters = [];
+
     /**
      * Saves response format data detected from XML tag "route".
      *
      * @param \SimpleXMLElement $info
+     * @throws ConfigurationException
      */
     public function __construct(\SimpleXMLElement $info)
     {
         parent::__construct($info);
-        $this->requestMethod = (string) $info["method"];
+        $this->setValidRequestMethod($info);
+        $this->setValidParameters($info);
+    }
+
+    /**
+     * Sets valid request method for route
+     *
+     * @param \SimpleXMLElement $info
+     * @throws ConfigurationException
+     */
+    private function setValidRequestMethod(\SimpleXMLElement $info): void
+    {
+        $method = (string) $info["method"];
+        if (!$method) {
+            return;
+        }
+        if ($case = Method::tryFrom(strtoupper($method))) {
+            $this->requestMethod = $case;
+        } else {
+            throw new ConfigurationException("Invalid request method: ".$method);
+        }
+    }
+    
+    /**
+     * Gets valid request method for current route
+     *
+     * @return ?Method
+     */
+    public function getValidRequestMethod(): ?Method
+    {
+        return $this->requestMethod;
+    }
+
+    /**
+     * Sets valid route/request parameters for route by name
+     *
+     * @param \SimpleXMLElement $info
+     * @throws ConfigurationException
+     */
+    private function setValidParameters(\SimpleXMLElement $info): void
+    {
         $parameters = $info->xpath("parameter");
         foreach ($parameters as $parameter) {
             $this->parameters[(string) $parameter["name"]] = new Parameter($parameter);
@@ -30,20 +70,9 @@ class Route extends \Lucinda\MVC\Application\Route
     }
     
     /**
-     * Gets valid request method for current route
-     *
-     * @return string|NULL
-     */
-    public function getValidRequestMethod(): ?string
-    {
-        return $this->requestMethod;
-    }
-    
-    /**
      * Gets validator for route/request parameter by its name for current route
      *
-     * @param string $name
-     * @return Parameter|NULL
+     * @return array
      */
     public function getValidParameters(): array
     {
