@@ -9,35 +9,12 @@ use Lucinda\MVC\Response;
  */
 abstract class Controller implements Runnable
 {
-    /**
-     * @var Application
-     */
-    protected $application;
-    
-    /**
-     * @var Request
-     */
-    protected $request;
-    
-    /**
-     * @var Session
-     */
-    protected $session;
-    
-    /**
-     * @var Cookies
-     */
-    protected $cookies;
-    
-    /**
-     * @var Response
-     */
-    protected $response;
-    
-    /**
-     * @var Attributes
-     */
-    protected $attributes;
+    protected Application $application;
+    protected Request $request;
+    protected Session $session;
+    protected Cookies $cookies;
+    protected Response $response;
+    protected Attributes $attributes;
         
     /**
      * Saves objects to be available in implemented run() methods.
