@@ -14,36 +14,12 @@ use Lucinda\MVC\Response;
  */
 abstract class End implements Runnable
 {
-    /**
-     * @var Attributes
-     */
-    protected $attributes;
-    
-    /**
-     * @var Application
-     */
-    protected $application;
-    
-    /**
-     * @var Request
-     */
-    protected $request;
-    
-    /**
-     * @var Session
-     */
-    protected $session;
-    
-    /**
-     * @var Cookies
-     */
-    protected $cookies;
-    
-    /**
-     * @var Response
-     */
-    protected $response;
-    
+    protected Attributes $attributes;
+    protected Application $application;
+    protected Request $request;
+    protected Session $session;
+    protected Cookies $cookies;
+    protected Response $response;
     
     /**
      * Saves objects to be available in implemented run() methods.
